EDX-5843 move contracts/client/interfaces to rest/interfaces
