+++ pkg/rest/interfaces/auth.go
