+++ pkg/rest/interfaces/authinjector.go
