+++ pkg/rest/interfaces/secretstoretoken.go
